+++ src/wrapper/atom.rs
@@ -4,7 +4,7 @@ use {NifResult, NifError};
 
 #[allow(dead_code)]
 pub unsafe fn make_atom(env: NIF_ENV, name: &str) -> NIF_TERM {
-    nif_interface::enif_make_atom_len(env, name.as_ptr() as *const u8, name.len() as usize)
+    nif_interface::enif_make_atom_len(env, name.as_ptr() as *const u8, name.len())
 }
 
 /// Get the contents of this atom as a string.
